# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Rute & Jadwal Service API",
        "description": "Dokumentasi API untuk Layanan Rute & Jadwal (Tugas 2 IAE)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local Docker Server"
        }
    ],
    "paths": {
        "/api/v1/schedules": {
            "get": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Mengambil daftar rute dan jadwal",
                "operationId": "e8a7715efdd9c1d13650dda726065bc4",
                "parameters": [
                    {
                        "name": "X-IAE-KEY",
                        "in": "header",
                        "description": "NIM Mahasiswa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data"
                    }
                }
            },
            "post": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Mendaftarkan jadwal armada baru",
                "operationId": "3a2079f068d082e4b1bdb4ac69b5cbfa",
                "parameters": [
                    {
                        "name": "X-IAE-KEY",
                        "in": "header",
                        "description": "NIM Mahasiswa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "route",
                                    "departure_time",
                                    "facilities",
                                    "price"
                                ],
                                "properties": {
                                    "route": {
                                        "type": "string",
                                        "example": "Bandung - Jakarta"
                                    },
                                    "departure_time": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-06-03 08:00:00"
                                    },
                                    "facilities": {
                                        "type": "string",
                                        "example": "AC, Reclining Seat, WiFi"
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 150000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Jadwal berhasil ditambahkan"
                    }
                }
            }
        },
        "/api/v1/schedules/{id}": {
            "get": {
                "tags": [
                    "Schedules"
                ],
                "summary": "Mengambil detail informasi jadwal spesifik",
                "operationId": "bd36ee741eebc9a8ba36a3914ac422ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Jadwal",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "X-IAE-KEY",
                        "in": "header",
                        "description": "NIM Mahasiswa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Schedules",
            "description": "Schedules"
        }
    ]
}